+++ SDK/AppCenterDistribute/Microsoft.AppCenter.Distribute.iOS/Distribute.cs
@@ -76,6 +76,8 @@ namespace Microsoft.AppCenter.Distribute
         static ReleaseAvailableCallback _releaseAvailableCallback;
 
         static WillExitAppCallback _willExitAppCallback;
+        
+        static NoReleaseAvailableCallback _noReleaseAvailableCallback;
 
         static void SetReleaseAvailableCallback(ReleaseAvailableCallback releaseAvailableCallback)
         {
@@ -96,6 +98,19 @@ namespace Microsoft.AppCenter.Distribute
             {
                 _willExitAppCallback = willExitAppCallback;
                 if (_delegate == null && _willExitAppCallback != null)
+                {
+                    _delegate = new Delegate();
+                    iOSDistribute.SetDelegate(_delegate);
+                }
+            }
+        }
+
+        static void SetNoReleaseAvailable(NoReleaseAvailableCallback noReleaseAvailable)
+        {
+            lock (typeof(Distribute))
+            {
+                _noReleaseAvailableCallback = noReleaseAvailable;
+                if (_delegate == null && _noReleaseAvailableCallback != null)
                 {
                     _delegate = new Delegate();
                     iOSDistribute.SetDelegate(_delegate);
@@ -172,6 +187,11 @@ namespace Microsoft.AppCenter.Distribute
             public override void WillExitApp(iOSDistribute distribute)
             {
                 _willExitAppCallback?.Invoke();
+            }
+
+            public override void OnNoReleaseAvailable(MSACDistribute distribute)
+            {
+                _noReleaseAvailableCallback?.Invoke();
             }
         }
     }