+++ SDK/AppCenterDistribute/Microsoft.AppCenter.Distribute/Distribute.cs
@@ -29,7 +29,11 @@ namespace Microsoft.AppCenter.Distribute
         {
         }
 
-        static void SetWillExitAppCallback(WillExitAppCallback willExitAppCallback)
+        static void SetWillExitAppCallback(WillExitAppCallback willExitAppCallback) 
+        {
+        }
+
+        static void SetNoReleaseAvailable(NoReleaseAvailableCallback noReleaseAvailable)
         {
         }